Massive Page Adding and Reorg
Reorganized all the pages in subfiles logically to make navigation
in coding easier. Also added all the pages that were wanted to the site.
Finally, slight modifications to stlye.css to make dropdowns more
logical and easier to replicate, along with some small things to keep
the nav bar centered and looking good.